# resources/openapi/mcp-v1.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TikTak.by MCP Analytics API",
    "version": "2.2.1",
...
  },
  "servers": [
    {
      "url": "https://tiktak.by/api/mcp/v1",
      "description": "Production"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
    "schemas": {
...
      "RedirectRow": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 42
          },
          "source_url": {
            "type": "string",
            "example": "/ru/old-path"
          },
          "target_url": {
            "type": "string",
            "example": "/ru/new-path"
          },
          "status_code": {
            "type": "integer",
            "enum": [
              301,
              302
            ],
            "example": 301
          },
          "is_active": {
            "type": "boolean",
            "example": true
          },
          "is_regex": {
            "type": "boolean",
            "example": false
          },
          "comment": {
            "type": "string",
            "nullable": true,
            "example": "SEO cleanup"
          },
          "hit_count": {
            "type": "integer",
            "example": 0
          },
          "last_hit_at": {
            "type": "string",
            "nullable": true,
            "example": "2026-05-26 12:00:00"
          }
        },
        "required": [
          "id",
          "source_url",
          "target_url",
          "status_code",
          "is_active",
          "is_regex",
          "hit_count"
        ]
      },
      "RedirectInput": {
        "type": "object",
        "required": [
          "source_url",
          "target_url"
        ],
        "properties": {
          "source_url": {
            "type": "string",
            "maxLength": 2000,
            "example": "/ru/old-path"
          },
          "target_url": {
            "type": "string",
            "maxLength": 2000,
            "example": "/ru/new-path"
          },
          "status_code": {
            "type": "integer",
            "enum": [
              301,
              302
            ],
            "default": 301
          },
          "is_active": {
            "type": "boolean",
            "default": true
          },
          "is_regex": {
            "type": "boolean",
            "default": false
          },
          "comment": {
            "type": "string",
            "nullable": true,
            "maxLength": 1000
          }
        }
      }
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Missing or invalid Bearer token."
      },
...
      "ValidationError": {
        "description": "Invalid query parameters.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                },
                "errors": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
...
      "RedirectListEnvelope": {
        "description": "Paginated list of redirect records.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "query": {
                  "type": "object"
                },
                "data": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RedirectRow"
                  }
                },
                "meta": {
                  "type": "object",
                  "properties": {
                    "total_rows": {
                      "type": "integer"
                    },
                    "page": {
                      "type": "integer"
                    },
                    "per_page": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "RedirectSingleResponse": {
        "description": "A single redirect record (create / update result).",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "$ref": "#/components/schemas/RedirectRow"
                },
                "meta": {
                  "type": "object",
                  "properties": {
                    "affected_rows": {
                      "type": "integer",
                      "example": 1
                    }
                  }
                }
              }
            }
          }
        }
      },
      "RedirectDeleteResponse": {
        "description": "Confirmation of redirect deletion.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "properties": {
                    "deleted_id": {
                      "type": "integer"
                    }
                  }
                },
                "meta": {
                  "type": "object",
                  "properties": {
                    "affected_rows": {
                      "type": "integer",
                      "example": 1
                    }
                  }
                }
              }
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/redirects": {
      "get": {
        "operationId": "tiktak_redirects_list",
        "summary": "List all redirect rules.",
        "description": "Returns paginated list of rows from the `redirects` table that drives the CheckRedirects middleware. Supports filtering and text search.",
        "parameters": [
          {
            "in": "query",
            "name": "is_active",
            "schema": {
              "type": "integer",
              "enum": [
                0,
                1
              ]
            },
            "description": "Filter by active status."
          },
          {
            "in": "query",
            "name": "is_regex",
            "schema": {
              "type": "integer",
              "enum": [
                0,
                1
              ]
            },
            "description": "Filter by regex flag."
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "LIKE match on source_url or target_url."
          },
          {
            "in": "query",
            "name": "per_page",
            "schema": {
              "type": "integer",
              "default": 100,
              "maximum": 500
            },
            "description": "Items per page."
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number."
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/RedirectListEnvelope"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "operationId": "tiktak_redirect_create",
        "summary": "Create a single redirect rule.",
        "description": "Inserts a new row into the `redirects` table and immediately clears the redirect cache. Returns 422 if source_url already exists.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RedirectInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/RedirectSingleResponse"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
...
    "/redirects/{id}": {
      "patch": {
        "operationId": "tiktak_redirect_update",
        "summary": "Partially update a redirect rule.",
        "description": "Updates only the provided fields. At least one field must be supplied. Clears cache on success.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Redirect record ID."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "source_url": {
                    "type": "string",
                    "maxLength": 2000
                  },
                  "target_url": {
                    "type": "string",
                    "maxLength": 2000
                  },
                  "status_code": {
                    "type": "integer",
                    "enum": [
                      301,
                      302
                    ]
                  },
                  "is_active": {
                    "type": "boolean"
                  },
                  "is_regex": {
                    "type": "boolean"
                  },
                  "comment": {
                    "type": "string",
                    "nullable": true,
                    "maxLength": 1000
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/RedirectSingleResponse"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "description": "Redirect not found."
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      },
      "delete": {
        "operationId": "tiktak_redirect_delete",
        "summary": "Delete a redirect rule.",
        "description": "Permanently removes the redirect and clears cache.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Redirect record ID."
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/RedirectDeleteResponse"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "description": "Redirect not found."
          }
        }
      }
    },
...
  },
...
}